Moved weather display to separate component
